将请求参数的枚举切换为rust枚举

+++ server/crates/service_hub/src/log/service/user_login.rs
@@ -110,7 +110,7 @@ impl<'a> UserLoginService<'a> {
             device: Set(Some(device)),
             system: Set(Some(system)),
             browser: Set(Some(browser)),
-            status: Set(data.status),
+            status: Set(data.status as i8),
             disabled: Set(UserLoginDisabledStatus::Enabled as i8),
             ..Default::default()
         };
